refactor(atom-svelte): build as plain TypeScript, drop svelte-package
With the read handles now backed by `createSubscriber` (and mounting via
`onDestroy`), the source no longer contains any runes — `Hooks.svelte.ts` is
plain TypeScript that simply imports from `svelte` / `svelte/reactivity`, just
like the Vue adapter imports from `vue`.

That removes the entire build divergence:

- Rename `Hooks.svelte.ts` -> `Hooks.ts`.
- Build with the shared `tsc -b && babel` pipeline and check with `tsc -b`,
  like the react/vue/solid siblings; drop `@sveltejs/package` and
  `svelte-check`.
- Add the package to `tsconfig.packages.json` and the root `tsconfig.json`
  path aliases, so it rejoins the tsc project graph.
- Remove `svelte.config.js`, the post-build `.d.ts` extension rewrite (the
  emitted `.ts` specifiers match every other package and are handled the same
  way at publish), and the `.svelte-kit` ignores. Revert the dprint and jsdocs
  config changes.

`@sveltejs/vite-plugin-svelte` is retained as a dev dependency only to compile
the `.svelte` test fixtures under vitest. Net root footprint is now just the
package's entries in `tsconfig.packages.json` and the root path aliases, the
same as any other package.
