+++ dprint.json
@@ -16,7 +16,6 @@
     "LLMS.md",
     "**/dist",
     "**/build",
-    "**/.svelte-kit",
     "**/docs",
     "**/coverage",
     "packages/**/CHANGELOG.md",
