+++ jsdocs.config.json
@@ -7,7 +7,6 @@
   "exclude": [
     "**/node_modules/**",
     "packages/tools/**",
-    "packages/atom/svelte/src/**",
     "packages/**/src/index.ts",
     "packages/**/src/*Generated.ts",
     "packages/**/src/internal/**",
