--- packages/atom/svelte/.gitignore
@@ -1,2 +1,0 @@
-# Generated by svelte-package
-.svelte-kit/
