--- packages/atom/svelte/svelte.config.js
@@ -1,5 +1,0 @@
-import { vitePreprocess } from "@sveltejs/vite-plugin-svelte"
-
-export default {
-  preprocess: vitePreprocess()
-}
