+++ tsconfig.json
@@ -51,6 +51,8 @@
       "@effect/atom-vue/*": ["./packages/atom/vue/src/*.ts"],
       "@effect/atom-solid": ["./packages/atom/solid/src/index.ts"],
       "@effect/atom-solid/*": ["./packages/atom/solid/src/*.ts"],
+      "@effect/atom-svelte": ["./packages/atom/svelte/src/index.ts"],
+      "@effect/atom-svelte/*": ["./packages/atom/svelte/src/*.ts"],
       "@effect/opentelemetry": ["./packages/opentelemetry/src/index.ts"],
       "@effect/opentelemetry/*": ["./packages/opentelemetry/src/*.ts"],
       "@effect/platform-browser": ["./packages/platform-browser/src/index.ts"],
